Refactor GitLab service into modular mixins pattern
- Extract GitLab service functionality into feature-based mixins
- Create GitLabMixinBase with common HTTP/GraphQL/auth methods
- Create GitLabReposMixin for repository operations
- Create GitLabBranchesMixin for branch operations
- Create GitLabPRsMixin for merge request operations
- Create GitLabFeaturesMixin for suggested tasks and microagents
- Create GitLabResolverMixin for comments and discussions
- Reduce main GitLabService from 926 lines to 67 lines
- Maintain identical public API and functionality
- Follow established GitHub service mixins pattern
