+++ openhands/integrations/gitlab/service/__init__.py
@@ -1,0 +1,15 @@
+# openhands/integrations/gitlab/service/__init__.py
+
+from .branches import GitLabBranchesMixin
+from .features import GitLabFeaturesMixin
+from .prs import GitLabPRsMixin
+from .repos import GitLabReposMixin
+from .resolver import GitLabResolverMixin
+
+__all__ = [
+    'GitLabBranchesMixin',
+    'GitLabFeaturesMixin',
+    'GitLabPRsMixin',
+    'GitLabReposMixin',
+    'GitLabResolverMixin',
+]
